+++ diagram-editor/frontend/edges/section-edge.tsx
@@ -1,6 +1,6 @@
 import { type EdgeProps, StepEdge } from '@xyflow/react';
 import { memo } from 'react';
-import type { DataEdge } from '.';
+import type { DataEdge } from './data-edge';
 
 export type SectionOutputData = {
   output: string;
@@ -13,13 +13,3 @@ export type SectionOutputEdgeProps = Exclude<EdgeProps<SectionEdge>, 'label'>;
 export const SectionOutputEdgeComp = memo((props: SectionOutputEdgeProps) => {
   return <StepEdge {...props} label={props.data.output.output} />;
 });
-
-export type SectionInputSlotData = {
-  type: 'sectionInput';
-  inputId: string;
-};
-
-export type SectionBufferSlotData = {
-  type: 'sectionBuffer';
-  inputId: string;
-};
